Renamed the "tokenise.corpus()" to "index()" and removed the older index.corpus() from the NAMESPACE.  Changed dfm() to use the "indexed" corpus if indexed.
Added settingsReset().

Made some notes and code changes to tokenize.R but PAUL STILL NEEDS TO WORK WITH THIS.

Removed inaugCorpus.RData, so this needs to be recreated.  And the inaugTexts.RData need to be checked (and should also be a named vector - the names are gone for some reason).