Error checking and argument handling for dictionary construction
Makes it possible to construct a list through dictionary(), without wrapping it in a list().

Adds a validator to the S4 class for dictionary, to ensure that all values are character, and better tests that all keys are supplied (as list names).

Adds constructor tests for dictionary for above conditions, and for file formats.

Closes #407