Add repeated and comparison tests for perfectly dissimilar matrix
On Ubuntu 16.04 LTS the last tests fail.

When I have just restarted R, the next to last tests fail (the same ones that are failing on CRAN).

This appears to be an instability when the matrixes are perfectly dissimilar, as in the edge case of the example matrix mt in the last examples.