Fix #1514
Corrects the problem where the keyword pattern was not displaying correctly at the top of the textplot_xray panel.  This was caused by the kwic attribute `keywords` not getting properly set.  Because this was different for characters versus dictionaries, I used a bit of hack to choose unique values (dictionary) or unique patterns (character vector).

The rest of the changes are just linting.